+++ CHANGELOG.md
@@ -19,6 +19,8 @@
 ### Changes to output format
 
 - Node `properties` objects now export keys in alphabetical order.
+- Omit default node property values from export. Set the `:json-omit-default-property-values` to
+  `nil` to restore old behavior.
 - `headline` nodes:
   - The added `tags-all` property has been moved from `"properties"` to the top level object.